requirements.txt
Django==2.0.2
django-bootstrap4==0.0.6
django-forms-bootstrap==3.1.0
django-select2-forms==2.0.2
django-sortedm2m==1.5.0
django-tables2==1.19.0
djangorestframework==3.7.7
mysqlclient==1.3.12
pkg-resources==0.0.0
psycopg2==2.7.3.2
pytz==2017.3
requirements==0.1
